requirements.txt
ctransformers==0.2.5
faiss-cpu==1.7.4
fastapi>=0.96.0
ipykernel>=6.23.1
langchain==0.0.315
pypdf==3.8.1
python-box==7.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
uvicorn>=0.22.0
llama_cpp_python==0.2.11
docx2txt==0.8
streamlit